feat: 공통 컴포넌트 (dropdown) 구현 (#99)
- 뷰 로직, 비지니스 로직 분리
- 스타일 수정
