+++ src/hooks/useDropdown.ts
@@ -1,0 +1,27 @@
+import { useState, useRef, useEffect } from 'react'
+
+export const useDropdown = (onClose?: () => void) => {
+  const [isOpen, setIsOpen] = useState(false)
+  const dropdownRef = useRef<HTMLDivElement>(null)
+
+  const handleToggle = () => setIsOpen(!isOpen)
+
+  const handleClickOutside = (event: MouseEvent) => {
+    if (
+      dropdownRef.current &&
+      !dropdownRef.current.contains(event.target as Node)
+    ) {
+      setIsOpen(false)
+      onClose?.()
+    }
+  }
+
+  useEffect(() => {
+    document.addEventListener('mousedown', handleClickOutside)
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside)
+    }
+  }, [])
+
+  return { isOpen, dropdownRef, handleToggle, setIsOpen }
+}
